Implement Keyboard and Timers
Keyboard is now allocated in main and is provided input from SDL. As recommended, (SDL) scan codes are used to ensure non-typical keyboard layouts are supported.

Additionally, timing has been setup, although it's not perfect. Most games seem to run way too fast, despite emulating 9 cycles per frame. How is this value determined? The CHIP-8 has a roughly 540Hz CPU and has a refresh rate of 60Hz. (540 / 60) = 9, so we want to execute 9 cycles per (drawing) frame.

Despite that knowledge, something is off! Pushing for now while I investigate

Finally, a mild refactor is in order as I begin to accumulate logic and variables in main. Likely move this to a "chip8.zig" file which will act as the binder to the CPU.